+++ src/build64.sh
@@ -8,5 +8,5 @@ echo ./gen mem.s
 ./gen mem.s
 fi
 
-echo gcc -g3 cImpls.c lisp.c cell.c mem64.c -o lisp64
-gcc -g3 cImpls.c lisp.c cell.c mem64.c -o lisp64
+echo gcc -g3 cImpls.c lisp.c cell.c mem64.c math.c -o lisp64
+gcc -g3 cImpls.c lisp.c cell.c mem64.c math.c -o lisp64
