+++ src/lisp.c
@@ -50,83 +50,6 @@ any EVAL(any x)
 
 
 static void gc(word c);
-uword getHeapSize(void);
-/* Prototypes */
-void *alloc(void*,size_t);
-any apply(any,any,bool,int,cell*);
-void argError(any,any) ;
-void atomError(any,any) ;
-void begString(void);
-void brkLoad(any);
-int bufNum(char[BITS/2],word);
-int bufSize(any);
-void bufString(any,char*);
-void bye(int) ;
-void pairError(any,any) ;
-any circ(any);
-word compare(any,any);
-any consIntern(any,any);
-any cons(any,any);
-any consName(uword,any);
-any consSym(any,uword);
-void newline(void);
-any endString(void);
-bool equal(any,any);
-void err(any,any,char*,...) ;
-any evExpr(any,any);
-any evList(any);
-word evNum(any,any);
-any evSym(any);
-void execError(char*) ;
-int firstByte(any);
-any get(any,any);
-int getByte(int*,uword*,any*);
-int getByte1(int*,uword*,any*);
-void getStdin(void);
-void giveup(char*) ;
-void heapAlloc(void);
-any intern(any,any[2]);
-any isIntern(any,any[2]);
-any load(any,int,any);
-any loadAll(any);
-any method(any);
-any mkChar(int);
-any mkSym(byte*);
-any mkStr(char*);
-any mkTxt(int);
-any name(any);
-void numError(any,any) ;
-any numToSym(any,int,int,int);
-void outName(any);
-void outNum(word);
-void outString(char*);
-void pack(any,int*,uword*,any*,cell*);
-int pathSize(any);
-void pathString(any,char*);
-void popInFiles(void);
-void popOutFiles(void);
-any popSym(int,uword,any,cell*);
-void prin(any);
-void print(any);
-void protError(any,any) ;
-void pushInFiles(inFrame*);
-void pushOutFiles(outFrame*);
-void put(any,any,any);
-void putByte(int,int*,uword*,any*,cell*);
-void putByte1(int,int*,uword*,any*);
-void putStdout(int);
-void rdOpen(any,any,inFrame*);
-any read1(int);
-void space(void);
-int symBytes(any);
-void symError(any,any) ;
-any symToNum(any,int,int,int);
-void undefined(any,any);
-void unwind (catchFrame*);
-void varError(any,any) ;
-void wrOpen(any,any,outFrame*);
-word xNum(any,any);
-any xSym(any);
 
 
 /* Globals */
@@ -490,53 +413,8 @@ any doHide(any ex)
    return Nil;
 }
 
+void makeError(any ex) {err(ex, NULL, "Not making");}
 
-
-
-
-static void makeError(any ex) {err(ex, NULL, "Not making");}
-
-// (link 'any ..) -> any
-any doLink(any x)
-{
-    any y;
-
-    if (!Env.make)
-    {
-        makeError(x);
-    }
-    x = cdr(x);
-    do
-    {
-        y = EVAL(car(x));
-        Env.make = &cdr(*Env.make = cons(y, Nil));
-    }
-    while (Nil != (x = cdr(x)));
-    return y;
-}
-
-// (make .. [(made 'lst ..)] .. [(link 'any ..)] ..) -> any
-any doMake(any x)
-{
-    any *make, *yoke;
-    cell c1;
-
-    Push(c1, Nil);
-    make = Env.make;
-    yoke = Env.yoke;
-    Env.make = Env.yoke = &data(c1);
-
-    while (Nil != (x = cdr(x)))
-    {
-        if (isCell(car(x)))
-        {
-            evList(car(x));
-        }
-    }
-    Env.yoke = yoke;
-    Env.make = make;
-    return Pop(c1);
-}
 
 ///////////////////////////////////////////////
 //               sym.c - END
@@ -1355,159 +1233,6 @@ any doPrin(any x)
 ///////////////////////////////////////////////
 
 
-/* Make number from symbol */
-any symToNum(any sym, int scl, int sep, int ign)
-{
-    unsigned c;
-    int i;
-    uword w;
-    bool sign, frac;
-    word n;
-    any s = sym;
-
-
-    if (!(c = getByte1(&i, &w, &s)))
-        return NULL;
-
-    while (c <= ' ')  /* Skip white space */
-    {
-        if (!(c = getByte(&i, &w, &s)))
-            return NULL;
-    }
-
-    sign = NO;
-    if (c == '+'  ||  c == '-' && (sign = YES))
-    {
-        if (!(c = getByte(&i, &w, &s)))
-            return NULL;
-    }
-
-    if ((c -= '0') > 9)
-        return NULL;
-
-    frac = NO;
-    n = c;
-    while ((c = getByte(&i, &w, &s))  &&  (!frac || scl))
-    {
-        if ((int)c == sep)
-        {
-            if (frac)
-                return NULL;
-            frac = YES;
-        }
-        else if ((int)c != ign)
-        {
-            if ((c -= '0') > 9)
-                return NULL;
-            n = n * 10 + c;
-            if (frac)
-                --scl;
-        }
-    }
-    if (c)
-    {
-        if ((c -= '0') > 9)
-            return NULL;
-        if (c >= 5)
-            n += 1;
-        while (c = getByte(&i, &w, &s))
-        {
-            if ((c -= '0') > 9)
-                return NULL;
-        }
-    }
-    if (frac)
-        while (--scl >= 0)
-            n *= 10;
-
-
-    any r = cons(Nil, Nil);
-    r->car = (any)n;
-    r->type.parts[0] = NUM;
-
-    return r;
-}
-
-any mkNum(word n)
-{
-    any r = cons(Nil, Nil);
-    r->car = (any)n;
-    r->type.parts[0] = NUM;
-    return r;
-}
-
-// (+ 'num ..) -> num
-any doAdd(any ex)
-{
-    any x, y;
-    uword n=0;
-
-    x = cdr(ex);
-    if (isNil(y = EVAL(car(x))))
-        return Nil;
-    NeedNum(ex,y);
-    n = unBox(y);
-    while (Nil != (x = cdr(x)))
-    {
-        if (isNil(y = EVAL(car(x))))
-            return Nil;
-        NeedNum(ex,y);
-        n += unBox(y);
-    }
-
-    any r = cons(Nil, Nil);
-    r->car = (any)n;
-    r->type.parts[0] = NUM;
-    return r;
-}
-
-any doSub(any ex)
-{
-    any x, y;
-    uword n=0;
-
-    x = cdr(ex);
-    if (isNil(y = EVAL(car(x))))
-        return Nil;
-    NeedNum(ex,y);
-    n = unBox(y);
-    while (Nil != (x = cdr(x)))
-    {
-        if (isNil(y = EVAL(car(x))))
-            return Nil;
-        NeedNum(ex,y);
-        n -= unBox(y);
-    }
-
-    any r = cons(Nil, Nil);
-    r->car = (any)n;
-    r->type.parts[0] = NUM;
-    return r;
-}
-
-any doMul(any ex)
-{
-    any x, y;
-    uword n=0;
-
-    x = cdr(ex);
-    if (isNil(y = EVAL(car(x))))
-        return Nil;
-    NeedNum(ex,y);
-    n = unBox(y);
-    while (Nil != (x = cdr(x)))
-    {
-        if (isNil(y = EVAL(car(x))))
-            return Nil;
-        NeedNum(ex,y);
-        n *= unBox(y);
-    }
-
-    any r = cons(Nil, Nil);
-    r->car = (any)n;
-    r->type.parts[0] = NUM;
-    return r;
-}
 
 ///////////////////////////////////////////////
 //               math.c END
